// lib/enums/VarType.ts
export enum VarType {
  char = 'char',
  char16_t = 'char16_t',
  char32_t = 'char32_t',
  wchar_t = 'wchar_t',
  unsigned_char = 'unsigned_char',

  short_int = 'short_int',
  int = 'int',
  long_int = 'long_int',
  long_long_int = 'long_long_int',

  unsigned_short_int = 'unsigned_short_int',
  unsigned_int = 'unsigned_int',
  unsigned_long_int = 'unsigned_long_int',
  unsigned_long_long_int = 'unsigned_long_long_int',

  int8_t = 'int8_t',
  int16_t = 'int16_t',
  int32_t = 'int32_t',
  int64_t = 'int64_t',

  uint8_t = 'uint8_t',
  uint16_t = 'uint16_t',
  uint32_t = 'uint32_t',
  uint64_t = 'uint64_t',

  float = 'float',
  double = 'double',

  bool = 'bool'
}